Move currency definitions to a separate module
